+++ plugins/python/src/main/kotlin/io/paddle/plugin/python/dependencies/index/distributions/ArchivePyDistributionInfo.kt
@@ -1,4 +1,4 @@
-package io.paddle.plugin.python.dependencies.index
+package io.paddle.plugin.python.dependencies.index.distributions
 
 import kotlinx.serialization.Serializable
